requirements.txt
psycopg2-binary>=2.9.9
python-docx>=0.8.11
ttkthemes>=3.2.2
python-dotenv>=1.0.0
tkinter
csv
os
requests>=2.31.0
